Correções comentadas na PR
