+++ src/components/confirmDialog/confirmDialog.types.ts
@@ -1,0 +1,11 @@
+export type ConfirmDialogProps = {
+    title: string;
+    description: string;
+    id: string;
+    queryKeyToInvalidate: string;
+    onConfirm: (id: string) => Promise<void>;
+    onCancel?: () => void;
+    open: boolean;
+    onOpenChange: (open: boolean) => void;
+    buttomLabel?: string;
+};
