+++ src/modules/bookUpsert/bookUpsert.tsx
@@ -8,12 +8,7 @@ import {
 } from "@/components/ui/dialog";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
-import { useEffect } from "react";
 import { SelectField } from "../home/components/select/select.";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { BookCreateValidator, Status } from "@/types/books.types";
-import { bookCreateSchema } from "@/modules/home/validators/createBook.validator";
 import { Checkbox } from "@/components/ui/checkbox";
 import { DatePicker } from "../home/components/datePicker/datePicker";
 import { useBookDialog } from "./useBookDialog";
@@ -29,19 +24,11 @@ import {
 } from "@/components/ui/form";
 import { Switch } from "@/components/ui/switch";
 import { Label } from "@/components/ui/label";
-import { SelectedBookshelf } from "../shelves/types/bookshelves.types";
-import { useQuery } from "@tanstack/react-query";
-import { BookshelfService } from "../shelves/services/booksshelves.service";
 import { Separator } from "@/components/ui/separator";
 import { useIsLoggedIn } from "@/stores/hooks/useAuth";
 import { BlurOverlay } from "@/components/blurOverlay/blurOverlay";
 import { ConfirmDialog } from "@/components/confirmDialog/confirmDialog";
-
-type CreateBookProps = {
-  bookData?: BookCreateValidator;
-  isBookFormOpen: boolean;
-  setIsBookFormOpen: (open: boolean) => void;
-};
+import { CreateBookProps } from "./bookUpsert.types";
 
 export function BookUpsert({
   bookData,
@@ -49,26 +36,9 @@ export function BookUpsert({
   setIsBookFormOpen,
 }: CreateBookProps) {
   const isLoggedIn = useIsLoggedIn();
-  const form = useForm<BookCreateValidator>({
-    resolver: zodResolver(bookCreateSchema),
-    defaultValues: {
-      title: "",
-      author: "",
-      pages: undefined,
-      readers: "",
-      start_date: null,
-      end_date: null,
-      gender: "",
-      image_url: "",
-      ...bookData,
-    },
-  });
 
-  const { reset, handleSubmit, control } = form;
-  const isEdit: boolean = Boolean(bookData && bookData.id);
   const {
     onSubmit,
-    createBook,
     isLoading,
     isAddToShelfEnabled,
     selected,
@@ -77,41 +47,20 @@ export function BookUpsert({
     setSelected,
     setSelectedShelfId,
     isDuplicateBookDialogOpen,
-    setIsDuplicateBookDialogOpen
-  } = useBookDialog({
+    setIsDuplicateBookDialogOpen,
+    form,
     reset,
+    handleSubmit,
+    control, 
+    checkboxes, 
+    isEdit, 
+    isLoadingBookShelfs, 
+    bookshelfOptions,
+    handleConfirmCreateBook,
+  } = useBookDialog({
     bookData,
     setIsBookFormOpen,
-    isEdit
   });
-  const checkboxes: { id: Status; label: string }[] = [
-    { id: "not_started", label: "Vou iniciar a leitura" },
-    { id: "reading", label: "Já iniciei a leitura" },
-    { id: "finished", label: "Terminei a Leitura" },
-  ];
-
-  useEffect(() => {
-    if (bookData) {
-      setSelected(bookData.status ?? null);
-    } else {
-      setSelected(null);
-    }
-  }, [bookData, setSelected]);
-
-  const { data: bookshelves = [], isLoading: isLoadingBookShelfs } = useQuery({
-    queryKey: ["bookshelves"],
-    queryFn: async () => {
-      const service = new BookshelfService();
-      return service.getAll();
-    },
-  });
-
-  const bookshelfOptions =
-    bookshelves?.map((shelf: SelectedBookshelf) => ({
-      label: shelf.name,
-      value: shelf.id,
-    })) ?? [];
-
   return (
     <>
       <ConfirmDialog
@@ -119,11 +68,7 @@ export function BookUpsert({
         onOpenChange={setIsDuplicateBookDialogOpen}
         title="Livro Duplicado"
         description="Um livro com este título já existe, deseja continuar?"
-        onConfirm={async () => {
-          setIsDuplicateBookDialogOpen(false);
-          const formData = form.getValues();
-          createBook.mutate(formData);
-        }}
+        onConfirm={() => handleConfirmCreateBook()}
         id="duplicate-book-warning"
         queryKeyToInvalidate="books"
         buttomLabel="Continuar"
