+++ src/modules/bookUpsert/bookUpsert.types.ts
@@ -1,0 +1,12 @@
+import { BookCreateValidator } from "@/types/books.types";
+
+export type CreateBookProps = {
+    bookData?: BookCreateValidator;
+    isBookFormOpen: boolean;
+    setIsBookFormOpen: (open: boolean) => void;
+};
+
+export type UseCreateBookDialog = {
+    bookData: BookCreateValidator | undefined;
+    setIsBookFormOpen: (open: boolean) => void;
+};
