+++ src/modules/bookUpsert/services/bookUpsert.service.ts
@@ -2,6 +2,7 @@ import { createClient } from "@/lib/supabase/client";
 import { BookCreateValidator } from "@/types/books.types";
 import { ErrorHandler, RepositoryError } from "@/services/errors/error";
 import { BookUpsertMapper } from "./mappers/bookUpsert.mapper";
+import { BookQueryBuilder } from "@/services/books/bookQuery.builder";
 
 export class BookUpsertService {
   private supabase = createClient();
@@ -74,5 +75,16 @@ export class BookUpsertService {
       ErrorHandler.log(normalizedError);
       throw normalizedError;
     }
+  }
+
+  async checkDuplicateBook(title: string): Promise<boolean> {
+    const supabase = createClient();
+    const duplicateCheckQuery = await new BookQueryBuilder(supabase)
+      .withSearchTerm(title)
+      .build();
+
+    const { data: booksWithSameTitle } = await duplicateCheckQuery;
+
+    return booksWithSameTitle?.length > 0;
   }
 }
